+++ Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/.idea/gradle.xml
@@ -10,6 +10,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/app/assets" />
           </set>
         </option>
         <option name="resolveExternalAnnotations" value="false" />
