+++ Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/app/src/main/java/com/example/smartwardrobeanalytics/App.kt
@@ -1,0 +1,10 @@
+package com.example.smartwardrobeanalytics
+
+import android.app.Application
+
+class App : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        Config.loadConfig(this)
+    }
+}
