+++ Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/app/src/main/java/com/example/smartwardrobeanalytics/Config.kt
@@ -5,11 +5,16 @@ import java.nio.charset.Charset
 
 object Config {
     lateinit var ONE_SIGNAL_APP_ID: String
+    lateinit var BASE_URL: String
 
     fun loadConfig(context: Context) {
         val jsonString = loadJSONFromAsset(context, "config.json")
+        if (jsonString == null) {
+            throw RuntimeException("Config file not found or is empty")
+        }
         val jsonObject = JSONObject(jsonString)
         ONE_SIGNAL_APP_ID = jsonObject.getString("ONESIGNAL_APP_ID")
+        BASE_URL = jsonObject.getString("BASE_URL")
     }
 
     private fun loadJSONFromAsset(context: Context, fileName: String): String? {
