+++ Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/app/src/main/java/com/example/smartwardrobeanalytics/activities/BrandListActivity.kt
@@ -3,7 +3,6 @@ package com.example.smartwardrobeanalytics.activities
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
-import android.view.MenuItem
 import android.widget.Button
 import android.widget.Toast
 import androidx.appcompat.app.ActionBarDrawerToggle
@@ -11,7 +10,6 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.smartwardrobeanalytics.LoginActivity
 import com.example.smartwardrobeanalytics.R
 import com.example.smartwardrobeanalytics.activities.createActivities.CreateBrandActivity
 import com.example.smartwardrobeanalytics.activities.editActivities.EditBrandActivity
