--- Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/app/src/main/java/com/example/smartwardrobeanalytics/assets/config.json
@@ -1,3 +1,0 @@
-{
-  "ONESIGNAL_APP_ID": "2cb74d5d-c9c8-4985-ae88-6ffd94aa43e9"
-}
