+++ Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/app/src/main/java/com/example/smartwardrobeanalytics/interfaces/iservices/IApiService.kt
@@ -1,4 +1,4 @@
-package com.example.smartwardrobeanalytics;
+package com.example.smartwardrobeanalytics.interfaces.iservices;
 
 import com.example.smartwardrobeanalytics.dtos.CollectionDto
 import com.example.smartwardrobeanalytics.dtos.CollectionResponse
