+++ Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/app/src/main/java/com/example/smartwardrobeanalytics/services/ItemServiceImpl.kt
@@ -22,7 +22,7 @@ class ItemServiceImpl {
         .build()
 
     private val retrofit = Retrofit.Builder()
-        .baseUrl("http://192.168.50.234:5002/api/")
+        .baseUrl(Config.BASE_URL)
         .addConverterFactory(GsonConverterFactory.create())
         .client(okHttpClient)
         .build()
