+++ Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/app/src/main/java/com/example/smartwardrobeanalytics/services/SeasonalStatisticsServiceImpl.kt
@@ -20,7 +20,7 @@ class SeasonalStatisticsServiceImpl {
         .build()
 
     private val retrofit = Retrofit.Builder()
-        .baseUrl("http://192.168.50.234:5002/api/")
+        .baseUrl(Config.BASE_URL)
         .addConverterFactory(GsonConverterFactory.create())
         .client(okHttpClient)
         .build()
