+++ Task5/apz-pzpi-21-4-molchan-stanislav-task5/SmartWardrobeAnalyticsMobile/settings.gradle.kts
@@ -24,3 +24,4 @@ dependencyResolutionManagement {
 
 rootProject.name = "SmartWardrobeAnalytics"
 include(":app")
+include(":app:assets")
